Move OperatorLogPointManager
