+++ packages/extension/src/commands/operatorLogPointManagement.ts
@@ -2,7 +2,7 @@ import { interfaces } from 'inversify';
 import * as vscode from 'vscode';
 import { ILogger } from '../logger';
 import OperatorLogPoint from '../operatorLogPoint';
-import { IOperatorLogPointManager } from '../operatorLogPoint/logPointManager';
+import { IOperatorLogPointManager } from '../operatorLogPoint/manager';
 import { Commands } from './commands';
 import registerCommand from './registerCommand';
 
@@ -10,7 +10,7 @@ export default function registerOperatorLogPointManagementCommands(
   context: vscode.ExtensionContext,
   container: interfaces.Container
 ): void {
-  const logPointManager = container.get<IOperatorLogPointManager>(IOperatorLogPointManager);
+  const manager = container.get<IOperatorLogPointManager>(IOperatorLogPointManager);
   const logger = container.get<ILogger>(ILogger);
 
   context.subscriptions.push(
@@ -18,7 +18,7 @@ export default function registerOperatorLogPointManagementCommands(
       if (typeof operatorLogPoint === 'string') {
         try {
           const parsed = OperatorLogPoint.parse(operatorLogPoint);
-          logPointManager.enable(parsed);
+          manager.enable(parsed);
         } catch (e) {
           logger.warn(
             'Extension',
@@ -26,7 +26,7 @@ export default function registerOperatorLogPointManagementCommands(
           );
         }
       } else {
-        logPointManager.enable(operatorLogPoint);
+        manager.enable(operatorLogPoint);
       }
     }),
 
@@ -34,7 +34,7 @@ export default function registerOperatorLogPointManagementCommands(
       if (typeof operatorLogPoint === 'string') {
         try {
           const parsed = OperatorLogPoint.parse(operatorLogPoint);
-          logPointManager.disable(parsed);
+          manager.disable(parsed);
         } catch (e) {
           logger.warn(
             'Extension',
@@ -42,7 +42,7 @@ export default function registerOperatorLogPointManagementCommands(
           );
         }
       } else {
-        logPointManager.disable(operatorLogPoint);
+        manager.disable(operatorLogPoint);
       }
     })
   );
