+++ packages/extension/src/decoration/operatorLogPointDecorationProvider/index.ts
@@ -10,7 +10,7 @@ import {
 } from 'vscode';
 import { DocumentDecorationProvider } from '..';
 import OperatorLogPoint from '../../operatorLogPoint';
-import { IOperatorLogPointManager } from '../../operatorLogPoint/logPointManager';
+import { IOperatorLogPointManager } from '../../operatorLogPoint/manager';
 import { IOperatorLogPointRecommendationEvent, IOperatorLogPointRecommender } from '../../operatorLogPoint/recommender';
 import { difference } from '../../util/map';
 import { IDisposable } from '../../util/types';
